# [Class] Make looking up a System.Type given a native Class instance faster